net: udp: Fix compilation after merging v5.10.235-rc1
../net/ipv4/udp.c:927:14: error: comparison of distinct pointer types ('typeof (datalen) *' (aka 'int *') and 'typeof (cork->gso_size) *' (aka 'unsigned short *')) [-Werror,-Wcompare-distinct-pointer-types]
  927 |                 if (hlen + min(datalen, cork->gso_size) > cork->fragsize) {
      |                            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ net/ipv4/udp.c
@@ -908,7 +908,7 @@ static int udp_send_skb(struct sk_buff *skb, struct flowi4 *fl4,
 	int is_udplite = IS_UDPLITE(sk);
 	int offset = skb_transport_offset(skb);
 	int len = skb->len - offset;
-	int datalen = len - sizeof(*uh);
+	unsigned short datalen = len - sizeof(*uh);
 	__wsum csum = 0;
 
 	/*
